refactor: reduce boilerplate, split chatsvc-api, typed responses, add tests
- Extract requireMessageAuthWithConfig() in auth-guards.ts (eliminates 4-line
  auth+config pattern repeated 13× across chatsvc functions)
- Extract handleSubstrateError() in auth-guards.ts (eliminates clearTokenCache
  pattern repeated 5× in substrate-api.ts and files-api.ts)
- Split chatsvc-api.ts (1,600+ lines) into focused sub-modules:
  - chatsvc-messaging.ts: send, edit, delete, threads, 1:1/group chat
  - chatsvc-activity.ts: activity feed (mentions, reactions, replies)
  - chatsvc-reactions.ts: add/remove emoji reactions
  - chatsvc-virtual.ts: saved messages, followed threads, save/unsave
  - chatsvc-readstatus.ts: consumption horizons, mark as read, unread
  - chatsvc-common.ts: shared date formatting utility
  - chatsvc-api.ts: barrel file re-exporting all sub-modules
- Add types/api-responses.ts with typed interfaces for raw API responses
  (RawChatsvcMessage, RawConversationResponse, RawAllFilesResponse, etc.)
- Apply typed interfaces to chatsvc-messaging, chatsvc-activity,
  chatsvc-readstatus, and files-api (reduces Record<string,unknown> casts)
- Add 83 new tests (200 total, was 117):
  - auth-guards.test.ts: handleSubstrateError
  - chatsvc-common.test.ts: formatHumanReadableDate
  - result.test.ts: ok/err/unwrap/unwrapOr/map/andThen
- Update AGENTS.md directory structure and ROADMAP.md (mark 5 tasks done)
